+++ docs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next release
+## 0.0.10
 
 * It is now [possible](configuration.md?id=OPENGRAPH_FALLBACK_IMAGE) to specify an Open Graph fallback image for `SearchMetaBase` subclasses.
 * Twitter fields have been removed from `SearchMetaBase` (and thus `Page`). These fields have for years not merited the prominence they have because Twitter has long supported OpenGraph metadata, and merit it even less now that Twitter is a dying website.